:muscle: Move `Plugin` impl into `service.ts` as private
`Plugin` is only used by `Service` and it is not used by other modules.
